wallet: introduce identity property
Use chain_code + address to generate an ID for a wallet.

This ID also serves as the filename of the wallet DB on the filesystem.
For backward compatibility, only use address to generate this ID for
BTC/ETH wallets so that we don't have to rename the file on the
filesystem.

Signed-off-by: Hunt Xu <taimanhui@protonmail.com>